integration/requirements.txt
pydantic>=2,<3
Pillow>=10,<12
